ffi: Split timeline into smaller modules
