+++ bindings/matrix-sdk-ffi/src/timeline/content.rs
@@ -1,7 +1,19 @@
+// Copyright 2023 The Matrix.org Foundation C.I.C.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 use std::{collections::HashMap, sync::Arc, time::Duration};
 
-use as_variant::as_variant;
-use eyeball_im::VectorDiff;
 use matrix_sdk::{
     attachment::{BaseAudioInfo, BaseFileInfo, BaseImageInfo, BaseThumbnailInfo, BaseVideoInfo},
     ruma::events::{
@@ -18,7 +30,6 @@ use matrix_sdk::{
                 LocationMessageEventContent as RumaLocationMessageEventContent,
                 MessageType as RumaMessageType,
                 NoticeMessageEventContent as RumaNoticeMessageEventContent,
-                RoomMessageEventContentWithoutRelation,
                 TextMessageEventContent as RumaTextMessageEventContent,
                 UnstableAudioDetailsContentBlock as RumaUnstableAudioDetailsContentBlock,
                 UnstableVoiceContentBlock as RumaUnstableVoiceContentBlock,
@@ -29,392 +40,15 @@ use matrix_sdk::{
         },
     },
 };
-use matrix_sdk_ui::timeline::{EventItemOrigin, PollResult, Profile, TimelineDetails};
+use matrix_sdk_ui::timeline::{PollResult, TimelineDetails};
 use ruma::{assign, serde::JsonObject, UInt};
 use tracing::{info, warn};
 
-use crate::{
-    error::{ClientError, MediaInfoError},
-    helpers::unwrap_or_clone_arc,
-    room::AssetType,
-};
-
-#[uniffi::export]
-pub fn media_source_from_url(url: String) -> Arc<MediaSource> {
-    Arc::new(MediaSource::Plain(url.into()))
-}
-
-#[uniffi::export]
-pub fn message_event_content_new(
-    msgtype: MessageType,
-) -> Result<Arc<RoomMessageEventContentWithoutRelation>, ClientError> {
-    Ok(Arc::new(RoomMessageEventContentWithoutRelation::new(msgtype.try_into()?)))
-}
-
-#[uniffi::export]
-pub fn message_event_content_from_markdown(
-    md: String,
-) -> Arc<RoomMessageEventContentWithoutRelation> {
-    Arc::new(RoomMessageEventContentWithoutRelation::new(RumaMessageType::text_markdown(md)))
-}
-
-#[uniffi::export]
-pub fn message_event_content_from_markdown_as_emote(
-    md: String,
-) -> Arc<RoomMessageEventContentWithoutRelation> {
-    Arc::new(RoomMessageEventContentWithoutRelation::new(RumaMessageType::emote_markdown(md)))
-}
-
-#[uniffi::export]
-pub fn message_event_content_from_html(
-    body: String,
-    html_body: String,
-) -> Arc<RoomMessageEventContentWithoutRelation> {
-    Arc::new(RoomMessageEventContentWithoutRelation::new(RumaMessageType::text_html(
-        body, html_body,
-    )))
-}
-
-#[uniffi::export]
-pub fn message_event_content_from_html_as_emote(
-    body: String,
-    html_body: String,
-) -> Arc<RoomMessageEventContentWithoutRelation> {
-    Arc::new(RoomMessageEventContentWithoutRelation::new(RumaMessageType::emote_html(
-        body, html_body,
-    )))
-}
-
-#[uniffi::export(callback_interface)]
-pub trait TimelineListener: Sync + Send {
-    fn on_update(&self, diff: Vec<Arc<TimelineDiff>>);
-}
+use super::ProfileDetails;
+use crate::{error::MediaInfoError, room::AssetType, utils::u64_to_uint};
 
 #[derive(Clone, uniffi::Object)]
-pub enum TimelineDiff {
-    Append { values: Vec<Arc<TimelineItem>> },
-    Clear,
-    PushFront { value: Arc<TimelineItem> },
-    PushBack { value: Arc<TimelineItem> },
-    PopFront,
-    PopBack,
-    Insert { index: usize, value: Arc<TimelineItem> },
-    Set { index: usize, value: Arc<TimelineItem> },
-    Remove { index: usize },
-    Truncate { length: usize },
-    Reset { values: Vec<Arc<TimelineItem>> },
-}
-
-impl TimelineDiff {
-    pub(crate) fn new(inner: VectorDiff<Arc<matrix_sdk_ui::timeline::TimelineItem>>) -> Self {
-        match inner {
-            VectorDiff::Append { values } => {
-                Self::Append { values: values.into_iter().map(TimelineItem::from_arc).collect() }
-            }
-            VectorDiff::Clear => Self::Clear,
-            VectorDiff::Insert { index, value } => {
-                Self::Insert { index, value: TimelineItem::from_arc(value) }
-            }
-            VectorDiff::Set { index, value } => {
-                Self::Set { index, value: TimelineItem::from_arc(value) }
-            }
-            VectorDiff::Truncate { length } => Self::Truncate { length },
-            VectorDiff::Remove { index } => Self::Remove { index },
-            VectorDiff::PushBack { value } => {
-                Self::PushBack { value: TimelineItem::from_arc(value) }
-            }
-            VectorDiff::PushFront { value } => {
-                Self::PushFront { value: TimelineItem::from_arc(value) }
-            }
-            VectorDiff::PopBack => Self::PopBack,
-            VectorDiff::PopFront => Self::PopFront,
-            VectorDiff::Reset { values } => {
-                warn!("Timeline subscriber lagged behind and was reset");
-                Self::Reset { values: values.into_iter().map(TimelineItem::from_arc).collect() }
-            }
-        }
-    }
-}
-
-#[uniffi::export]
-impl TimelineDiff {
-    pub fn change(&self) -> TimelineChange {
-        match self {
-            Self::Append { .. } => TimelineChange::Append,
-            Self::Insert { .. } => TimelineChange::Insert,
-            Self::Set { .. } => TimelineChange::Set,
-            Self::Remove { .. } => TimelineChange::Remove,
-            Self::PushBack { .. } => TimelineChange::PushBack,
-            Self::PushFront { .. } => TimelineChange::PushFront,
-            Self::PopBack => TimelineChange::PopBack,
-            Self::PopFront => TimelineChange::PopFront,
-            Self::Clear => TimelineChange::Clear,
-            Self::Truncate { .. } => TimelineChange::Truncate,
-            Self::Reset { .. } => TimelineChange::Reset,
-        }
-    }
-
-    pub fn append(self: Arc<Self>) -> Option<Vec<Arc<TimelineItem>>> {
-        let this = unwrap_or_clone_arc(self);
-        as_variant!(this, Self::Append { values } => values)
-    }
-
-    pub fn insert(self: Arc<Self>) -> Option<InsertData> {
-        let this = unwrap_or_clone_arc(self);
-        as_variant!(this, Self::Insert { index, value } => {
-            InsertData { index: index.try_into().unwrap(), item: value }
-        })
-    }
-
-    pub fn set(self: Arc<Self>) -> Option<SetData> {
-        let this = unwrap_or_clone_arc(self);
-        as_variant!(this, Self::Set { index, value } => {
-            SetData { index: index.try_into().unwrap(), item: value }
-        })
-    }
-
-    pub fn remove(&self) -> Option<u32> {
-        as_variant!(self, Self::Remove { index } => (*index).try_into().unwrap())
-    }
-
-    pub fn push_back(self: Arc<Self>) -> Option<Arc<TimelineItem>> {
-        let this = unwrap_or_clone_arc(self);
-        as_variant!(this, Self::PushBack { value } => value)
-    }
-
-    pub fn push_front(self: Arc<Self>) -> Option<Arc<TimelineItem>> {
-        let this = unwrap_or_clone_arc(self);
-        as_variant!(this, Self::PushFront { value } => value)
-    }
-
-    pub fn reset(self: Arc<Self>) -> Option<Vec<Arc<TimelineItem>>> {
-        let this = unwrap_or_clone_arc(self);
-        as_variant!(this, Self::Reset { values } => values)
-    }
-}
-
-#[derive(uniffi::Record)]
-pub struct InsertData {
-    pub index: u32,
-    pub item: Arc<TimelineItem>,
-}
-
-#[derive(uniffi::Record)]
-pub struct SetData {
-    pub index: u32,
-    pub item: Arc<TimelineItem>,
-}
-
-#[derive(Clone, Copy, uniffi::Enum)]
-pub enum TimelineChange {
-    Append,
-    Clear,
-    Insert,
-    Set,
-    Remove,
-    PushBack,
-    PushFront,
-    PopBack,
-    PopFront,
-    Truncate,
-    Reset,
-}
-
-#[repr(transparent)]
-#[derive(Clone, uniffi::Object)]
-pub struct TimelineItem(pub(crate) matrix_sdk_ui::timeline::TimelineItem);
-
-impl TimelineItem {
-    pub(crate) fn from_arc(arc: Arc<matrix_sdk_ui::timeline::TimelineItem>) -> Arc<Self> {
-        // SAFETY: This is valid because Self is a repr(transparent) wrapper
-        //         around the other Timeline type.
-        unsafe { Arc::from_raw(Arc::into_raw(arc) as _) }
-    }
-}
-
-#[uniffi::export]
-impl TimelineItem {
-    pub fn as_event(self: Arc<Self>) -> Option<Arc<EventTimelineItem>> {
-        let event_item = self.0.as_event()?;
-        Some(Arc::new(EventTimelineItem(event_item.clone())))
-    }
-
-    pub fn as_virtual(self: Arc<Self>) -> Option<VirtualTimelineItem> {
-        use matrix_sdk_ui::timeline::VirtualTimelineItem as VItem;
-        match self.0.as_virtual()? {
-            VItem::DayDivider(ts) => Some(VirtualTimelineItem::DayDivider { ts: ts.0.into() }),
-            VItem::ReadMarker => Some(VirtualTimelineItem::ReadMarker),
-        }
-    }
-
-    pub fn unique_id(&self) -> u64 {
-        self.0.unique_id()
-    }
-
-    pub fn fmt_debug(&self) -> String {
-        format!("{:#?}", self.0)
-    }
-}
-
-/// This type represents the “send state” of a local event timeline item.
-#[derive(Clone, uniffi::Enum)]
-pub enum EventSendState {
-    /// The local event has not been sent yet.
-    NotSentYet,
-    /// The local event has been sent to the server, but unsuccessfully: The
-    /// sending has failed.
-    SendingFailed { error: String },
-    /// Sending has been cancelled because an earlier event in the
-    /// message-sending queue failed.
-    Cancelled,
-    /// The local event has been sent successfully to the server.
-    Sent { event_id: String },
-}
-
-impl From<&matrix_sdk_ui::timeline::EventSendState> for EventSendState {
-    fn from(value: &matrix_sdk_ui::timeline::EventSendState) -> Self {
-        use matrix_sdk_ui::timeline::EventSendState::*;
-
-        match value {
-            NotSentYet => Self::NotSentYet,
-            SendingFailed { error } => Self::SendingFailed { error: error.to_string() },
-            Cancelled => Self::Cancelled,
-            Sent { event_id } => Self::Sent { event_id: event_id.to_string() },
-        }
-    }
-}
-
-#[derive(uniffi::Object)]
-pub struct EventTimelineItem(pub(crate) matrix_sdk_ui::timeline::EventTimelineItem);
-
-#[uniffi::export]
-impl EventTimelineItem {
-    pub fn is_local(&self) -> bool {
-        self.0.is_local_echo()
-    }
-
-    pub fn is_remote(&self) -> bool {
-        !self.0.is_local_echo()
-    }
-
-    pub fn transaction_id(&self) -> Option<String> {
-        self.0.transaction_id().map(ToString::to_string)
-    }
-
-    pub fn event_id(&self) -> Option<String> {
-        self.0.event_id().map(ToString::to_string)
-    }
-
-    pub fn sender(&self) -> String {
-        self.0.sender().to_string()
-    }
-
-    pub fn sender_profile(&self) -> ProfileDetails {
-        self.0.sender_profile().into()
-    }
-
-    pub fn is_own(&self) -> bool {
-        self.0.is_own()
-    }
-
-    pub fn is_editable(&self) -> bool {
-        self.0.is_editable()
-    }
-
-    pub fn content(&self) -> Arc<TimelineItemContent> {
-        Arc::new(TimelineItemContent(self.0.content().clone()))
-    }
-
-    pub fn timestamp(&self) -> u64 {
-        self.0.timestamp().0.into()
-    }
-
-    pub fn reactions(&self) -> Vec<Reaction> {
-        self.0
-            .reactions()
-            .iter()
-            .map(|(k, v)| Reaction {
-                key: k.to_owned(),
-                count: v.len().try_into().unwrap(),
-                senders: v
-                    .senders()
-                    .map(|v| ReactionSenderData {
-                        sender_id: v.sender_id.to_string(),
-                        timestamp: v.timestamp.0.into(),
-                    })
-                    .collect(),
-            })
-            .collect()
-    }
-
-    pub fn debug_info(&self) -> EventTimelineItemDebugInfo {
-        EventTimelineItemDebugInfo {
-            model: format!("{:#?}", self.0),
-            original_json: self.0.original_json().map(|raw| raw.json().get().to_owned()),
-            latest_edit_json: self.0.latest_edit_json().map(|raw| raw.json().get().to_owned()),
-        }
-    }
-
-    pub fn local_send_state(&self) -> Option<EventSendState> {
-        self.0.send_state().map(Into::into)
-    }
-
-    pub fn read_receipts(&self) -> HashMap<String, Receipt> {
-        self.0.read_receipts().iter().map(|(k, v)| (k.to_string(), v.clone().into())).collect()
-    }
-
-    pub fn origin(&self) -> Option<EventItemOrigin> {
-        self.0.origin()
-    }
-
-    pub fn can_be_replied_to(&self) -> bool {
-        self.0.can_be_replied_to()
-    }
-}
-
-#[derive(uniffi::Record)]
-pub struct Receipt {
-    pub timestamp: Option<u64>,
-}
-
-impl From<ruma::events::receipt::Receipt> for Receipt {
-    fn from(value: ruma::events::receipt::Receipt) -> Self {
-        Receipt { timestamp: value.ts.map(|ts| ts.0.into()) }
-    }
-}
-
-#[derive(uniffi::Record)]
-pub struct EventTimelineItemDebugInfo {
-    model: String,
-    original_json: Option<String>,
-    latest_edit_json: Option<String>,
-}
-
-#[derive(uniffi::Enum)]
-pub enum ProfileDetails {
-    Unavailable,
-    Pending,
-    Ready { display_name: Option<String>, display_name_ambiguous: bool, avatar_url: Option<String> },
-    Error { message: String },
-}
-
-impl From<&TimelineDetails<Profile>> for ProfileDetails {
-    fn from(details: &TimelineDetails<Profile>) -> Self {
-        match details {
-            TimelineDetails::Unavailable => Self::Unavailable,
-            TimelineDetails::Pending => Self::Pending,
-            TimelineDetails::Ready(profile) => Self::Ready {
-                display_name: profile.display_name.clone(),
-                display_name_ambiguous: profile.display_name_ambiguous,
-                avatar_url: profile.avatar_url.as_ref().map(ToString::to_string),
-            },
-            TimelineDetails::Error(e) => Self::Error { message: e.to_string() },
-        }
-    }
-}
-
-#[derive(Clone, uniffi::Object)]
-pub struct TimelineItemContent(matrix_sdk_ui::timeline::TimelineItemContent);
+pub struct TimelineItemContent(pub(crate) matrix_sdk_ui::timeline::TimelineItemContent);
 
 #[uniffi::export]
 impl TimelineItemContent {
@@ -1078,13 +712,6 @@ impl From<&RumaFileInfo> for FileInfo {
     }
 }
 
-pub(crate) fn u64_to_uint(u: u64) -> UInt {
-    UInt::new(u).unwrap_or_else(|| {
-        warn!("u64 -> UInt conversion overflowed, falling back to UInt::MAX");
-        UInt::MAX
-    })
-}
-
 #[derive(uniffi::Record)]
 pub struct InReplyToDetails {
     event_id: String,
@@ -1163,20 +790,6 @@ pub struct Reaction {
 pub struct ReactionSenderData {
     pub sender_id: String,
     pub timestamp: u64,
-}
-
-/// A [`TimelineItem`](super::TimelineItem) that doesn't correspond to an event.
-#[derive(uniffi::Enum)]
-pub enum VirtualTimelineItem {
-    /// A divider between messages of two days.
-    DayDivider {
-        /// A timestamp in milliseconds since Unix Epoch on that day in local
-        /// time.
-        ts: u64,
-    },
-
-    /// The user's own read marker.
-    ReadMarker,
 }
 
 #[derive(Clone, uniffi::Enum)]
