+++ bindings/matrix-sdk-ffi/src/utils.rs
@@ -1,0 +1,23 @@
+// Copyright 2023 The Matrix.org Foundation C.I.C.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use ruma::UInt;
+use tracing::warn;
+
+pub(crate) fn u64_to_uint(u: u64) -> UInt {
+    UInt::new(u).unwrap_or_else(|| {
+        warn!("u64 -> UInt conversion overflowed, falling back to UInt::MAX");
+        UInt::MAX
+    })
+}
